Drop a search feature
Solr is hard to maintain. It consumes much memory, fails to reindex.
I need more memory to implement a redis-necessary feature.

Because the search feature is not used frequently and I think it is
not so necessary, I close it for now.

If you have some usecase to require the feature, please submit an issue
on GitHub. https://github.com/k0kubun/githubranking/issues